Almost done
Spent about three hours getting the notifications up and running.
Worked through most of the problems except for one.  Trying to
eliminate a problem where the app forces open when the notification
goes off.  It's been reduced to a screen flicker, but still is bad
enough to notice.  Something about the activity coming forward then
closing..